Necessary changes to the nondeterministic parses to compensate for the change in the way the mapping between states of the model and the rows of the transition matrix are handled.
- All tests are green.
- Some comments are now a bit wrong.

Next up: Correct comments.


Former-commit-id: 610c0282b26e11cbe02804fe3daa9b241ed352b1

+++ src/parser/AtomicPropositionLabelingParser.cpp
@@ -126,7 +126,7 @@ namespace storm {
 			buf = trimWhitespaces(buf);
 
 			uint_fast64_t state = 0;
-			uint_fast64_t lastState = -1;
+			uint_fast64_t lastState = (uint_fast64_t)-1;
 			cnt = 0;
 
 			// Now parse the assignments of labels to nodes.
@@ -137,7 +137,7 @@ namespace storm {
 				state = checked_strtol(buf, &buf);
 
 				// If the state has already been read or skipped once there might be a problem with the file (doubled lines, or blocks).
-				if(state <= lastState && lastState != -1) {
+				if(state <= lastState && lastState != (uint_fast64_t)-1) {
 					LOG4CPLUS_ERROR(logger, "Wrong file format in (" << filename << "). State " << state << " was found but has already been read or skipped previously.");
 					throw storm::exceptions::WrongFormatException() << "State " << state << " was found but has already been read or skipped previously.";
 				}
